Add files via upload
The Color Hunt Gaming Project is an application written in Java intended for the desktop, designed as a brain teaser intended to sharpen the mind. The game involves clicking on the correct color name, as displayed on the screen. The difficult part is that the colors of the text may differ from the color written, putting a challenge on the player's focus and thought process. The player loses points for incorrect clicks and must reach the target score within the time of constraints. Development using the Java programming language and desktop interface framework for the project.